# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adaptive Psychological Monitoring Platform",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/risk-assessments/latest": {
      "get": {
        "tags": [
          "Risk Assessment"
        ],
        "summary": "Get Latest Risk Assessment",
        "operationId": "get_latest_risk_assessment_api_v1_risk_assessments_latest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/risk-assessments": {
      "get": {
        "tags": [
          "Risk Assessment"
        ],
        "summary": "List Risk Assessments",
        "operationId": "list_risk_assessments_api_v1_risk_assessments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RiskAssessmentOut"
                  },
                  "type": "array",
                  "title": "Response List Risk Assessments Api V1 Risk Assessments Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/risk-assessments/{assessment_id}": {
      "get": {
        "tags": [
          "Risk Assessment"
        ],
        "summary": "Get Risk Assessment",
        "operationId": "get_risk_assessment_api_v1_risk_assessments__assessment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assessment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/me": {
      "get": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Get My Recommendations",
        "operationId": "get_my_recommendations_api_v1_recommendations_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PatientRecommendationOut"
                  },
                  "type": "array",
                  "title": "Response Get My Recommendations Api V1 Recommendations Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/recommendations/{patient_recommendation_id}/viewed": {
      "post": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Mark Recommendation Viewed",
        "operationId": "mark_recommendation_viewed_api_v1_recommendations__patient_recommendation_id__viewed_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Recommendation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{patient_recommendation_id}/feedback": {
      "post": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Submit Recommendation Feedback",
        "operationId": "submit_recommendation_feedback_api_v1_recommendations__patient_recommendation_id__feedback_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Recommendation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationFeedback"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/daily": {
      "get": {
        "tags": [
          "Reports & Analytics"
        ],
        "summary": "Get Daily Report",
        "operationId": "get_daily_report_api_v1_reports_daily_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/weekly": {
      "get": {
        "tags": [
          "Reports & Analytics"
        ],
        "summary": "Get Weekly Report",
        "operationId": "get_weekly_report_api_v1_reports_weekly_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/monthly": {
      "get": {
        "tags": [
          "Reports & Analytics"
        ],
        "summary": "Get Monthly Report",
        "operationId": "get_monthly_report_api_v1_reports_monthly_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/risk-progression": {
      "get": {
        "tags": [
          "Reports & Analytics"
        ],
        "summary": "Get Risk Progression",
        "operationId": "get_risk_progression_api_v1_reports_risk_progression_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RiskProgressionPoint"
                  },
                  "type": "array",
                  "title": "Response Get Risk Progression Api V1 Reports Risk Progression Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientRecommendationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recommendation": {
            "$ref": "#/components/schemas/RecommendationOut"
          },
          "delivered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Delivered At"
          },
          "is_viewed": {
            "type": "boolean",
            "title": "Is Viewed"
          },
          "is_helpful_feedback": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Helpful Feedback"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recommendation",
          "delivered_at",
          "is_viewed"
        ],
        "title": "PatientRecommendationOut"
      },
      "RecommendationCategory": {
        "type": "string",
        "enum": [
          "breathing_exercise",
          "relaxation",
          "sleep_tip",
          "stress_management",
          "motivational",
          "educational",
          "professional_help"
        ],
        "title": "RecommendationCategory"
      },
...
      "RecommendationFeedback": {
        "properties": {
          "is_helpful": {
            "type": "boolean",
            "title": "Is Helpful"
          }
        },
        "type": "object",
        "required": [
          "is_helpful"
        ],
        "title": "RecommendationFeedback"
      },
      "RecommendationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "$ref": "#/components/schemas/RecommendationCategory"
          },
          "title_ar": {
            "type": "string",
            "title": "Title Ar"
          },
          "content_ar": {
            "type": "string",
            "title": "Content Ar"
          },
          "media_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "title_ar",
          "content_ar"
        ],
        "title": "RecommendationOut"
      },
...
      "ReportOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "report_type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "summary_ar": {
            "type": "string",
            "title": "Summary Ar"
          },
          "metrics_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics Json"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_type",
          "period_start",
          "period_end",
          "summary_ar",
          "metrics_json",
          "generated_at"
        ],
        "title": "ReportOut"
      },
      "ReportType": {
        "type": "string",
        "enum": [
          "daily",
          "weekly",
          "monthly"
        ],
        "title": "ReportType"
      },
      "RiskAssessmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "interview_session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Interview Session Id"
          },
          "risk_level": {
            "type": "integer",
            "title": "Risk Level"
          },
          "anxiety_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Anxiety Score"
          },
          "stress_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Stress Score"
          },
          "sadness_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Sadness Score"
          },
          "burnout_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Burnout Score"
          },
          "sleep_quality_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Sleep Quality Score"
          },
          "adherence_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Adherence Score"
          },
          "composite_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Composite Score"
          },
          "explanation_ar": {
            "type": "string",
            "title": "Explanation Ar"
          },
          "explanation_factors_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanation Factors Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "interview_session_id",
          "risk_level",
          "anxiety_score",
          "stress_score",
          "sadness_score",
          "burnout_score",
          "sleep_quality_score",
          "adherence_score",
          "composite_score",
          "explanation_ar",
          "explanation_factors_json",
          "created_at"
        ],
        "title": "RiskAssessmentOut"
      },
...
      "RiskProgressionPoint": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "risk_level": {
            "type": "integer",
            "title": "Risk Level"
          },
          "composite_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Composite Score"
          }
        },
        "type": "object",
        "required": [
          "date",
          "risk_level",
          "composite_score"
        ],
        "title": "RiskProgressionPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}